+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Fixes LD saving events inconsistencies - ash
 - Relay no longer caches GraphQL responses which contain `errors` field - ash
 - Changes order of Favorites tabs and re-centers Favorites content when a new tab is selected - ashley
+- Adds pressReleaseUrl to show moreInfo - kieran
 
 ### 1.8.26